Add fixes from code review, #5790
