+++ src/main/java/org/jabref/gui/util/ThemeLoader.java
@@ -39,7 +39,7 @@ public class ThemeLoader {
 
     public static final String MAIN_CSS = "Base.css";
     public static final String DARK_CSS = "Dark.css";
-    public static String CUSTOM_CSS = "";
+    private static String CUSTOM_CSS = "";
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ThemeLoader.class);
     private final Optional<URL> additionalCssToLoad;
