+++ src/main/resources/l10n/JabRef_en.properties
@@ -456,12 +456,6 @@ Importing=Importing
 
 Importing\ in\ unknown\ format=Importing in unknown format
 
-Import\ CSS=Import CSS
-
-Import\ custom\ CSS\ file=Import custom CSS file
-
-Theme\ change\ to\ a\ custom\ theme.=Theme change to a custom theme.
-
 Include\ subgroups\:\ When\ selected,\ view\ entries\ contained\ in\ this\ group\ or\ its\ subgroups=Include subgroups: When selected, view entries contained in this group or its subgroups
 
 Independent\ group\:\ When\ selected,\ view\ only\ this\ group's\ entries=Independent group: When selected, view only this group's entries
@@ -2126,14 +2120,13 @@ Field\ cannot\ be\ empty.\ Please\ enter\ a\ name.=Field cannot be empty. Please
 Shared\ database=Shared database
 Lookup=Lookup
 
-Import\ theme=Import theme
-
-Export\ theme=Export theme
-
 Export\ theme\ as\ CSS=Export theme as CSS
-
 Could\ not\ export\ theme=Could not export theme
-
 Path\ to\ theme=Path to theme
-
 Theme\ name=Theme name
+Theme\ changed\ to\ a\ custom\ theme.=Theme changed to a custom theme.
+Import\ theme=Import theme
+Export\ Theme=Export Theme
+
+Import\ CSS=Import CSS
+Import\ custom\ CSS\ file=Import custom CSS file
